refactor: createRewriteRuleProxyServiceImpl を別ファイルに切り出し
PRレビューコメントに従い、background.ts から
RewriteRuleProxyServiceImpl.ts に実装を切り出し。
background.ts の責務を軽減し、保守性を向上。
