+++ host-frontend-root/frontend-src-root/src/frameworks-and-drivers/messaging/RewriteRuleProxyServiceImpl.ts
@@ -1,0 +1,28 @@
+import { IRewriteRuleRepository } from 'src/application/ports/IRewriteRuleRepository';
+import { container } from 'src/frameworks-and-drivers/di/container';
+import { IRewriteRuleProxyService } from 'src/frameworks-and-drivers/messaging/RewriteRuleProxyService';
+
+/**
+ * RewriteRuleProxyService の実装を作成
+ * container を使用してリポジトリを解決し、ルールを取得する
+ *
+ * Background Script専用: このファイルは container に依存するため、
+ * Content Script からはインポートしないこと
+ */
+export function createRewriteRuleProxyServiceImpl(): IRewriteRuleProxyService {
+  return {
+    async getAllRules() {
+      const repository = container.resolve<IRewriteRuleRepository>('IRewriteRuleRepository');
+      const rules = await repository.getAll();
+
+      return rules.toArray().map((rule) => ({
+        id: rule.id,
+        oldString: rule.oldString,
+        newString: rule.newString,
+        urlPattern: rule.urlPattern,
+        isRegex: rule.isRegex,
+        isActive: rule.isActive,
+      }));
+    },
+  };
+}
